Initial configuration of dynamodb in development mode.
Initial configuration of Swagger.

Creation of the user domain class.

Creation of User Repository, Service, and Resource classes for basic CRUD implementation.

Creation of the CustomUserRepository class to implement the data creation and loading features in the user table.

Creating exception classes to centralize the handling and generation of error messages through the ResourceExceptionHandler class.